# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-LIVRE",
        "version": "1.0"
    },
    "paths": {
...
        "/api/matieres": {
            "get": {
                "tags": [
                    "matieres"
                ],
                "summary": "liste des matières",
                "description": "liste des matières",
                "operationId": "matieres_list",
                "responses": {
                    "200": {
                        "description": "Affiche la liste des matières"
                    }
                }
            },
            "post": {
                "tags": [
                    "matieres"
                ],
                "summary": "Création d'une matière",
                "description": "Création d'une matière",
                "operationId": "matieres",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "designation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "designation": "EPS"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/matieres/{id}": {
            "get": {
                "tags": [
                    "matieres"
                ],
                "summary": "Atteindre une matiere",
                "description": "Obtenir les données d'une matiere",
                "operationId": "matieres_one",
                "responses": {
                    "200": {
                        "description": "Affichage d'une matiere"
                    }
                }
            },
            "put": {
                "tags": [
                    "matieres"
                ],
                "summary": "Mise à jour d'une matière",
                "description": "Mise à jour d'une matière",
                "operationId": "matieres_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "designation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "designation": "EPS"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "matieres"
                ],
                "summary": "Supprimer une matiere",
                "description": "Supprimer les données d'une matiere",
                "operationId": "matieres_delete",
                "responses": {
                    "200": {
                        "description": "Suppression d'une matiere"
                    }
                }
            }
        },
...
    }
}